Add cast to RAND_MAX

+++ src/arch_host.h
@@ -48,8 +48,8 @@ inline static float arch_random_float(unsigned long long seed, unsigned long lon
   }
   /* Use Box Muller algorithm to get a float from a normal distribution */
   const float two_pi = 2.0f * M_PI;
-  const float u1 = (float) rand() / RAND_MAX;
-  const float u2 = (float) rand() / RAND_MAX;
+  const float u1 = (float) rand() / (float)RAND_MAX;
+  const float u2 = (float) rand() / (float)RAND_MAX;
   const float factor = stdev * sqrtf (-2.0f * logf (u1));
   const float trig_arg = two_pi * u2;
   /* Box Muller algorithm produces two random normally distributed floats, z0 and z1 */
